// app/features/flights/types.ts
export interface FlightRecord {
  id: string;
  month: string;
  code: string;
  no: string;
  flt: string;
  departure: string;
  dow: string;
  sector: string;
  fltReturn: string;
  returnDate: string;
  dowReturn: string;
  sectorReturn: string;
  d: string;
  n: string;
  bc: string;
  ey: string;
  pakej: string;
  hariDiSaudi: string;
  remark: string;
}